Cast return value to int

+++ 0x10-variadic_functions/0-sum_them_all.c
@@ -28,5 +28,5 @@ int sum_them_all(const unsigned int n, ...)
 	}
 	/* Clean up */
 	va_end(argList);
-	return (sum);
+	return ((int)sum);
 }
